+++ src/components/ReleaseNotesLink/index.js
@@ -1,1 +1,2 @@
+export * from './ReleaseNotesLink';
 export { default } from './ReleaseNotesLink';